bugfix: Clean all temporary files before today

+++ internal/tmp.go
@@ -20,6 +20,8 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"strconv"
+	"strings"
 
 	"github.com/version-fox/vfox/internal/util"
 )
@@ -30,7 +32,27 @@ type Temp struct {
 }
 
 func (t *Temp) Remove() {
-	_ = os.RemoveAll(t.CurProcessPath)
+	dir, err := os.ReadDir(t.dirPath)
+	if err == nil {
+		_ = os.RemoveAll(t.CurProcessPath)
+		for _, file := range dir {
+			if !file.IsDir() {
+				continue
+			}
+			names := strings.SplitN(file.Name(), "-", 2)
+			if len(names) != 2 {
+				continue
+			}
+			timestamp := names[0]
+			i, err := strconv.ParseInt(timestamp, 10, 64)
+			if err != nil {
+				continue
+			}
+			if util.IsBeforeToday(i) {
+				_ = os.Remove(filepath.Join(t.dirPath, file.Name()))
+			}
+		}
+	}
 }
 
 func NewTemp(dirPath string, pid int) (*Temp, error) {
